+++ src/lib.rs
@@ -20,7 +20,7 @@
 //!
 //! ```toml
 //! [dev-dependencies]
-//! readme-sync = "0.2.0"
+//! readme-sync = "0.2.1"
 //! ```
 //!
 //! Then add an integration test using the necessary readme and docs modifiers,
@@ -191,7 +191,7 @@
 //! Or, in any Rust release, you can disable all `readme-sync` dependencies with:
 //! ```toml
 //! [dev-dependencies.readme-sync]
-//! version = "0.2.0"
+//! version = "0.2.1"
 //! default-features = false
 //! ```
 //!