+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.2.1] - 2021-12-19
 ### Added
 - `CMarkDocs::map_links` function to fix links including intra-doc links.
 - `{CMarkDocs, CMarkReadme}::{into_data, iter_events}`
@@ -50,7 +52,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Markdown transformations.
 - Synchronization status diagnostics.
 
-[Unreleased]: https://github.com/zheland/readme-sync/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/zheland/readme-sync/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/zheland/readme-sync/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/zheland/readme-sync/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/zheland/readme-sync/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/zheland/readme-sync/compare/v0.0.1...v0.1.0